requirements.txt
# File: requirements.txt

paddlepaddle==2.6.1
paddleocr==2.7.0.3
Pillow
customtkinter==5.2.2
numpy==1.26.4
PyMuPDF
fpdf2
argostranslate==1.9.0
pyside6